Move some SQLite functionality into SQLite module
This includes:

  - The custom Date/UUID representations
  - The JSON functions
  - The `rowid` column
  - The temporary trigger helpers
  - FTS5

Not included (yet):

  - Our insert statement still supports some SQLite-only functionality (the
    `OR` clause).
  - Need to audit the scalar/aggregate function helpers for any
    SQLite-specific functions.
