+++ Sources/StructuredQueriesCore/ScalarFunctions.swift
@@ -401,7 +401,10 @@ public struct QueryFunction<QueryValue>: QueryExpression {
   let name: QueryFragment
   let arguments: [QueryFragment]
 
-  init<each Argument: QueryExpression>(_ name: QueryFragment, _ arguments: repeat each Argument) {
+  package init<each Argument: QueryExpression>(
+    _ name: QueryFragment,
+    _ arguments: repeat each Argument
+  ) {
     self.name = name
     self.arguments = Array(repeat each arguments)
   }
