+++ Sources/StructuredQueriesSQLiteCore/Collation+SQLite.swift
