+++ Sources/StructuredQueriesSQLiteCore/FTS5.swift
