+++ Sources/StructuredQueriesSQLiteCore/QueryRepresentable/Date+JulianDay.swift
