+++ Sources/StructuredQueriesSQLiteCore/QueryRepresentable/UUID+Bytes.swift
