+++ Sources/StructuredQueriesSQLiteCore/QueryRepresentable/UUID+Uppercased.swift
