+++ Sources/StructuredQueriesSQLiteCore/Triggers.swift
